npcs.ex and map.ex broken up
